Fix pet characteristic enum columns
The fields for pet sex, size, species, and status were not saving correctly because the model defined enums, which by default are backed by an integer column, but the database columns were strings.

This updates the database column type and adds validations for the enum fields.

It also alphabetizes the schema file, which is now [the default in Rails 8.1](https://guides.rubyonrails.org/8_1_release_notes.html#active-record-notable-changes).